+++ src/index.js
@@ -1,11 +1,20 @@
 const { OrchestrationMachine, WorkflowError } = require('./orchestration-machine')
 const { MemoryRepository } = require('./memory-repository')
 const { HybridPipeline, PipelineError } = require('./hybrid-pipeline')
+const { LoopControl, LoopControlError } = require('./loop-control')
+const { OutputBoundary, OutputBoundaryError } = require('./output-boundary')
+const { CollectiveControl, CollectiveControlError } = require('./collective-control')
 
 module.exports = {
   MemoryRepository,
   OrchestrationMachine,
   WorkflowError,
   HybridPipeline,
-  PipelineError
+  PipelineError,
+  LoopControl,
+  LoopControlError,
+  OutputBoundary,
+  OutputBoundaryError,
+  CollectiveControl,
+  CollectiveControlError
 }